ensure default detectors as a single function
