+++ src/sentry/workflow_engine/migration_helpers/issue_alert_migration.py
@@ -1,21 +1,13 @@
 import logging
 from typing import Any
 
-from django.db import router, transaction
-from rest_framework import status
-
-from sentry.api.exceptions import SentryAPIException
 from sentry.constants import ObjectStatus
 from sentry.grouping.grouptype import ErrorGroupType
-from sentry.issues import grouptype
-from sentry.locks import locks
-from sentry.models.project import Project
 from sentry.models.rule import Rule, RuleSource
 from sentry.models.rulesnooze import RuleSnooze
 from sentry.rules.conditions.event_frequency import EventUniqueUserFrequencyConditionWithConditions
 from sentry.rules.conditions.every_event import EveryEventCondition
 from sentry.rules.processing.processor import split_conditions_and_filters
-from sentry.utils.locking import UnableToAcquireLock
 from sentry.workflow_engine.migration_helpers.issue_alert_conditions import (
     create_event_unique_user_frequency_condition_with_conditions,
     translate_to_data_condition,
@@ -38,71 +30,11 @@ from sentry.workflow_engine.models.data_condition import (
     Condition,
     enforce_data_condition_json_schema,
 )
-from sentry.workflow_engine.types import ERROR_DETECTOR_NAME, ISSUE_STREAM_DETECTOR_NAME
-from sentry.workflow_engine.typings.grouptype import IssueStreamGroupType
+from sentry.workflow_engine.types import ERROR_DETECTOR_NAME
 
 logger = logging.getLogger(__name__)
 
 SKIPPED_CONDITIONS = [Condition.EVERY_EVENT]
-VALID_DEFAULT_DETECTOR_TYPES = [ErrorGroupType.slug, IssueStreamGroupType.slug]
-
-
-class UnableToAcquireLockApiError(SentryAPIException):
-    status_code = status.HTTP_400_BAD_REQUEST
-    code = "unable_to_acquire_lock"
-    message = "Unable to acquire lock for issue alert migration."
-
-
-def ensure_default_detector(project: Project, type: str) -> Detector:
-    """
-    Ensure that the default error detector exists for a project.
-    If the Detector doesn't already exist, we try to acquire a lock to avoid double-creating,
-    and UnableToAcquireLockApiError if that fails.
-    """
-    group_type = grouptype.registry.get_by_slug(type)
-    if not group_type:
-        raise ValueError(f"Group type {type} not registered")
-    slug = group_type.slug
-    if slug not in VALID_DEFAULT_DETECTOR_TYPES:
-        raise ValueError(f"Invalid default detector type: {slug}")
-
-    # If it already exists, life is simple and we can return immediately.
-    # If there happen to be duplicates, we prefer the oldest.
-    existing = Detector.objects.filter(type=slug, project=project).order_by("id").first()
-    if existing:
-        return existing
-
-    # If we may need to create it, we acquire a lock to avoid double-creating.
-    # There isn't a unique constraint on the detector, so we can't rely on get_or_create
-    # to avoid duplicates.
-    # However, by only locking during the one-time creation, the window for a race condition is small.
-    lock = locks.get(
-        f"workflow-engine-project-{slug}-detector:{project.id}",
-        duration=2,
-        name=f"workflow_engine_default_{slug}_detector",
-    )
-    try:
-        with (
-            # Creation should be fast, so it's worth blocking a little rather
-            # than failing a request.
-            lock.blocking_acquire(initial_delay=0.1, timeout=3),
-            transaction.atomic(router.db_for_write(Detector)),
-        ):
-            detector, _ = Detector.objects.get_or_create(
-                type=slug,
-                project=project,
-                defaults={
-                    "config": {},
-                    "name": (
-                        ERROR_DETECTOR_NAME
-                        if slug == ErrorGroupType.slug
-                        else ISSUE_STREAM_DETECTOR_NAME
-                    ),
-                },
-            )
-            return detector
-    except UnableToAcquireLock:
-        raise UnableToAcquireLockApiError
 
 
 class IssueAlertMigrator:
